core/metrics/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gaia-metrics"
version = "0.1.0"
description = "Gaia metrics and logging utilities"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gaia_metrics"]

